Move http status code to other class
