--- app/src/main/java/com/codesoom/assignment/Constant.java
@@ -1,16 +1,0 @@
-package com.codesoom.assignment;
-
-/**
- * Contains constants used in application.
- */
-public class Constant {
-    static final int PORT = 8000;
-    static final int BACKLOG = 0;
-
-    public class HttpStatusCode {
-        public static final int OK = 200;
-        public static final int CREATED = 201;
-        public static final int NO_CONTENT = 204;
-        public static final int NOT_FOUND = 404;
-    }
-}
