+++ app/src/main/java/com/codesoom/assignment/HttpStatusCode.java
@@ -1,0 +1,8 @@
+package com.codesoom.assignment;
+
+public class HttpStatusCode {
+    public static final int OK = 200;
+    public static final int CREATED = 201;
+    public static final int NO_CONTENT = 204;
+    public static final int NOT_FOUND = 404;
+}
