+++ app/src/main/java/com/codesoom/assignment/ServerConfiguration.java
@@ -1,0 +1,9 @@
+package com.codesoom.assignment;
+
+/**
+ * Contains constants used in application.
+ */
+public class ServerConfiguration {
+    static final int PORT = 8000;
+    static final int BACKLOG = 0;
+}
